% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_newDS.R
\name{DESeq2_newDS}
\alias{DESeq2_newDS}
\title{Differential expression analysis on the server-side}
\usage{
DESeq2_newDS(vars, set, test, fitType, sfType, reduced, contrast)
}
\arguments{
\item{vars}{variables for the model}

\item{set}{\code{SummarizedExperiment} object}

\item{test}{"Wald" or "LRT"}

\item{fitType}{"parametric", "local", or "mean"}

\item{sfType}{"ratio", "poscounts", or "iterate"}

\item{reduced}{test="LRT"}

\item{contrast}{model matrix contrast}
}
\description{
Performs the DGE based on
the Negative Binomial distribution
}
\details{
This function perform a DGE analysis
for \code{SummarizedExperiment} input
}
\author{
S.Avesani
}
